Move FENData class to separate module
